feat(Analytics & Dashboard): integrate analytics data fetching and enhance dashboard API responses

+++ src/app/api/organization/join-org/delete-request/route.ts
@@ -18,6 +18,20 @@ export async function POST(req: NextRequest) {
 
   if (roleError) return roleError;
 
+  const invite = await prismaInstance.invite.findFirst({
+    where: {
+      id: inviteId,
+      organizationId: selectedOrg,
+    },
+    select: {
+      email: true,
+    },
+  });
+
+  if (!invite) {
+    return NextResponse.json({ error: "no invite found" }, { status: 404 });
+  }
+
   await prismaInstance.invite.deleteMany({
     where: {
       id: inviteId,
@@ -25,5 +39,13 @@ export async function POST(req: NextRequest) {
     },
   });
 
-  return NextResponse.json({ success: true }, { status: 200 });
+  const createdActivity = await prismaInstance.activity.create({
+    data: {
+      description: `Declined join request from ${invite.email}`,
+      userId: auth.user.id,
+      organizationId: selectedOrg,
+    },
+  });
+
+  return NextResponse.json({ success: true, createdActivity }, { status: 200 });
 }
